move initializers out of resnet.py
